requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-community
langchain-openai
faiss-cpu
gitpython
openai
python-dotenv
tree-sitter
sentence-transformers
torch
streamlit
requests
pydantic
